[5.8] Add renderable functionality to Illuminate\Notifications\Messages\MailMessage
This PR adds the Renderable contract to the MailMessage class, allowing easier previewing of mail notifications in the browser. 

In your controller:
```php
return (new FooNotification())->toMail('example@example.com');
```

This is parallel with the current functionality of the Mailable class, and obviates the need to search for code snippets to accomplish the same for Notifications. This addresses Issue #24073.