Remove Nexmo routing
After seeing https://github.com/laravel/framework/pull/29679 I noticed this was left over from when we removed the Nexmo notification channel from the core. After this change people will need to implement the routeNotificationForNexmo method themselves as they do for the Slack driver.

This is a breaking change for anyone using the Nexmo channel so this should be in the upgrade guide.